Break out as much common logic as possible
